add repositories

+++ backend/src/main/kotlin/com/bftcom/backend/entity/Work.kt
@@ -3,5 +3,5 @@ package com.bftcom.backend.entity
 data class Work(
 	override var id: Long = 0,
 	val title: String,
-	val genreId: Int
+	val genreId: Long
 ) : Entity
