+++ backend/src/main/kotlin/com/bftcom/backend/entity/WorkAuthor.kt
@@ -2,6 +2,6 @@ package com.bftcom.backend.entity
 
 data class WorkAuthor(
 	override var id: Long = 0,
-	val workId: Int,
-	val authorId: Int
+	val workId: Long,
+	val authorId: Long
 ) : Entity
